+++ client/mock/request.go
@@ -1,0 +1,62 @@
+package mock
+
+import (
+	"github.com/micro/go-micro/client"
+	"github.com/micro/go-micro/codec"
+)
+
+type rpcRequest struct {
+	service     string
+	endpoint    string
+	contentType string
+	codec       codec.Codec
+	body        interface{}
+	opts        client.RequestOptions
+}
+
+func newRequest(service, endpoint string, request interface{}, reqOpts ...client.RequestOption) client.Request {
+	var opts client.RequestOptions
+
+	for _, o := range reqOpts {
+		o(&opts)
+	}
+
+	contentType := client.DefaultContentType
+
+	// set the content-type specified
+	if len(opts.ContentType) > 0 {
+		contentType = opts.ContentType
+	}
+
+	return &rpcRequest{
+		service:     service,
+		endpoint:    endpoint,
+		body:        request,
+		contentType: contentType,
+		opts:        opts,
+	}
+}
+
+func (r *rpcRequest) ContentType() string {
+	return r.contentType
+}
+
+func (r *rpcRequest) Service() string {
+	return r.service
+}
+
+func (r *rpcRequest) Endpoint() string {
+	return r.endpoint
+}
+
+func (r *rpcRequest) Body() interface{} {
+	return r.body
+}
+
+func (r *rpcRequest) Codec() codec.Writer {
+	return r.codec
+}
+
+func (r *rpcRequest) Stream() bool {
+	return r.opts.Stream
+}
