+++ client/rpc/request.go
@@ -1,6 +1,7 @@
-package client
+package rpc
 
 import (
+	"github.com/micro/go-micro/client"
 	"github.com/micro/go-micro/codec"
 )
 
@@ -10,11 +11,11 @@ type rpcRequest struct {
 	contentType string
 	codec       codec.Codec
 	body        interface{}
-	opts        RequestOptions
+	opts        client.RequestOptions
 }
 
-func newRequest(service, endpoint string, request interface{}, contentType string, reqOpts ...RequestOption) Request {
-	var opts RequestOptions
+func newRequest(service, endpoint string, request interface{}, contentType string, reqOpts ...client.RequestOption) client.Request {
+	var opts client.RequestOptions
 
 	for _, o := range reqOpts {
 		o(&opts)
