+++ client/rpc/rpc_test.go
@@ -1,10 +1,11 @@
-package client
+package rpc
 
 import (
 	"context"
 	"fmt"
 	"testing"
 
+	"github.com/micro/go-micro/client"
 	"github.com/micro/go-micro/errors"
 	"github.com/micro/go-micro/registry"
 	"github.com/micro/go-micro/registry/mock"
@@ -17,8 +18,8 @@ func TestCallAddress(t *testing.T) {
 	endpoint := "Test.Endpoint"
 	address := "10.1.10.1:8080"
 
-	wrap := func(cf CallFunc) CallFunc {
-		return func(ctx context.Context, addr string, req Request, rsp interface{}, opts CallOptions) error {
+	wrap := func(cf client.CallFunc) client.CallFunc {
+		return func(ctx context.Context, addr string, req client.Request, rsp interface{}, opts client.CallOptions) error {
 			called = true
 
 			if req.Service() != service {
@@ -40,15 +41,15 @@ func TestCallAddress(t *testing.T) {
 
 	r := mock.NewRegistry()
 	c := NewClient(
-		Registry(r),
-		WrapCall(wrap),
+		client.Registry(r),
+		client.WrapCall(wrap),
 	)
 	c.Options().Selector.Init(selector.Registry(r))
 
 	req := c.NewRequest(service, endpoint, nil)
 
 	// test calling remote address
-	if err := c.Call(context.Background(), req, nil, WithAddress(address)); err != nil {
+	if err := c.Call(context.Background(), req, nil, client.WithAddress(address)); err != nil {
 		t.Fatal("call with address error", err)
 	}
 
@@ -65,8 +66,8 @@ func TestCallRetry(t *testing.T) {
 
 	var called int
 
-	wrap := func(cf CallFunc) CallFunc {
-		return func(ctx context.Context, addr string, req Request, rsp interface{}, opts CallOptions) error {
+	wrap := func(cf client.CallFunc) client.CallFunc {
+		return func(ctx context.Context, addr string, req client.Request, rsp interface{}, opts client.CallOptions) error {
 			called++
 			if called == 1 {
 				return errors.InternalServerError("test.error", "retry request")
@@ -79,15 +80,15 @@ func TestCallRetry(t *testing.T) {
 
 	r := mock.NewRegistry()
 	c := NewClient(
-		Registry(r),
-		WrapCall(wrap),
+		client.Registry(r),
+		client.WrapCall(wrap),
 	)
 	c.Options().Selector.Init(selector.Registry(r))
 
 	req := c.NewRequest(service, endpoint, nil)
 
 	// test calling remote address
-	if err := c.Call(context.Background(), req, nil, WithAddress(address)); err != nil {
+	if err := c.Call(context.Background(), req, nil, client.WithAddress(address)); err != nil {
 		t.Fatal("call with address error", err)
 	}
 
@@ -106,8 +107,8 @@ func TestCallWrapper(t *testing.T) {
 	port := 8080
 	address := "10.1.10.1:8080"
 
-	wrap := func(cf CallFunc) CallFunc {
-		return func(ctx context.Context, addr string, req Request, rsp interface{}, opts CallOptions) error {
+	wrap := func(cf client.CallFunc) client.CallFunc {
+		return func(ctx context.Context, addr string, req client.Request, rsp interface{}, opts client.CallOptions) error {
 			called = true
 
 			if req.Service() != service {
@@ -129,8 +130,8 @@ func TestCallWrapper(t *testing.T) {
 
 	r := mock.NewRegistry()
 	c := NewClient(
-		Registry(r),
-		WrapCall(wrap),
+		client.Registry(r),
+		client.WrapCall(wrap),
 	)
 	c.Options().Selector.Init(selector.Registry(r))
